requirements.txt
pandas
requests
beautifulsoup4
regex
nltk
openpyxl
python-dotenv
